新增CRUD

+++ elim/src/main/java/com/example/elim/dao/BusinessDao.java
@@ -3,5 +3,5 @@ package com.example.elim.dao;
 import com.example.elim.model.Business;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BusinessDao extends CrudRepository<Business, Long> {
+public interface BusinessDao extends CrudRepository<Business, Integer> {
 }
